debt - do not depend from services to core (status bar)
