+++ src/vs/workbench/services/statusbar/browser/statusbar.ts
@@ -9,14 +9,107 @@ import { ThemeColor } from 'vs/platform/theme/common/themeService';
 import { Event } from 'vs/base/common/event';
 import { Command } from 'vs/editor/common/languages';
 import { IMarkdownString } from 'vs/base/common/htmlContent';
-import { IStatusbarEntryLocation } from 'vs/workbench/browser/parts/statusbar/statusbarModel';
 import { ColorIdentifier } from 'vs/platform/theme/common/colorRegistry';
 
 export const IStatusbarService = createDecorator<IStatusbarService>('statusbarService');
 
+export interface IStatusbarService {
+
+	readonly _serviceBrand: undefined;
+
+	/**
+	 * An event that is triggered when an entry's visibility is changed.
+	 */
+	readonly onDidChangeEntryVisibility: Event<{ id: string; visible: boolean }>;
+
+	/**
+	 * Adds an entry to the statusbar with the given alignment and priority. Use the returned accessor
+	 * to update or remove the statusbar entry.
+	 *
+	 * @param id identifier of the entry is needed to allow users to hide entries via settings
+	 * @param alignment either LEFT or RIGHT side in the status bar
+	 * @param priority items get arranged from highest priority to lowest priority from left to right
+	 * in their respective alignment slot
+	 */
+	addEntry(entry: IStatusbarEntry, id: string, alignment: StatusbarAlignment, priority?: number): IStatusbarEntryAccessor;
+
+	/**
+	 * Adds an entry to the statusbar with the given alignment relative to another entry. Use the returned
+	 * accessor to update or remove the statusbar entry.
+	 *
+	 * @param id identifier of the entry is needed to allow users to hide entries via settings
+	 * @param alignment either LEFT or RIGHT side in the status bar
+	 * @param location a reference to another entry to position relative to
+	 */
+	addEntry(entry: IStatusbarEntry, id: string, alignment: StatusbarAlignment, location?: IStatusbarEntryLocation): IStatusbarEntryAccessor;
+
+	/**
+	 * Return if an entry is visible or not.
+	 */
+	isEntryVisible(id: string): boolean;
+
+	/**
+	 * Allows to update an entry's visibility with the provided ID.
+	 */
+	updateEntryVisibility(id: string, visible: boolean): void;
+
+	/**
+	 * Focused the status bar. If one of the status bar entries was focused, focuses it directly.
+	 */
+	focus(preserveEntryFocus?: boolean): void;
+
+	/**
+	 * Focuses the next status bar entry. If none focused, focuses the first.
+	 */
+	focusNextEntry(): void;
+
+	/**
+	 * Focuses the previous status bar entry. If none focused, focuses the last.
+	 */
+	focusPreviousEntry(): void;
+
+	/**
+	 *	Returns true if a status bar entry is focused.
+	 */
+	isEntryFocused(): boolean;
+
+	/**
+	 * Temporarily override statusbar style.
+	 */
+	overrideStyle(style: IStatusbarStyleOverride): IDisposable;
+}
+
 export const enum StatusbarAlignment {
 	LEFT,
 	RIGHT
+}
+
+export interface IStatusbarEntryLocation {
+
+	/**
+	 * The identifier of another status bar entry to
+	 * position relative to.
+	 */
+	id: string;
+
+	/**
+	 * The alignment of the status bar entry relative
+	 * to the referenced entry.
+	 */
+	alignment: StatusbarAlignment;
+
+	/**
+	 * Whether to move the entry close to the location
+	 * so that it appears as if both this entry and
+	 * the location belong to each other.
+	 */
+	compact?: boolean;
+}
+
+export function isStatusbarEntryLocation(thing: unknown): thing is IStatusbarEntryLocation {
+	const candidate = thing as IStatusbarEntryLocation | undefined;
+
+	return typeof candidate?.id === 'string' && typeof candidate.alignment === 'number';
 }
 
 export const ShowTooltipCommand: Command = {
@@ -92,72 +185,6 @@ export interface IStatusbarEntry {
 	 * Will enable a spinning icon in front of the text to indicate progress.
 	 */
 	readonly showProgress?: boolean;
-}
-
-export interface IStatusbarService {
-
-	readonly _serviceBrand: undefined;
-
-	/**
-	 * An event that is triggered when an entry's visibility is changed.
-	 */
-	readonly onDidChangeEntryVisibility: Event<{ id: string; visible: boolean }>;
-
-	/**
-	 * Adds an entry to the statusbar with the given alignment and priority. Use the returned accessor
-	 * to update or remove the statusbar entry.
-	 *
-	 * @param id identifier of the entry is needed to allow users to hide entries via settings
-	 * @param alignment either LEFT or RIGHT side in the status bar
-	 * @param priority items get arranged from highest priority to lowest priority from left to right
-	 * in their respective alignment slot
-	 */
-	addEntry(entry: IStatusbarEntry, id: string, alignment: StatusbarAlignment, priority?: number): IStatusbarEntryAccessor;
-
-	/**
-	 * Adds an entry to the statusbar with the given alignment relative to another entry. Use the returned
-	 * accessor to update or remove the statusbar entry.
-	 *
-	 * @param id identifier of the entry is needed to allow users to hide entries via settings
-	 * @param alignment either LEFT or RIGHT side in the status bar
-	 * @param location a reference to another entry to position relative to
-	 */
-	addEntry(entry: IStatusbarEntry, id: string, alignment: StatusbarAlignment, location?: IStatusbarEntryLocation): IStatusbarEntryAccessor;
-
-	/**
-	 * Return if an entry is visible or not.
-	 */
-	isEntryVisible(id: string): boolean;
-
-	/**
-	 * Allows to update an entry's visibility with the provided ID.
-	 */
-	updateEntryVisibility(id: string, visible: boolean): void;
-
-	/**
-	 * Focused the status bar. If one of the status bar entries was focused, focuses it directly.
-	 */
-	focus(preserveEntryFocus?: boolean): void;
-
-	/**
-	 * Focuses the next status bar entry. If none focused, focuses the first.
-	 */
-	focusNextEntry(): void;
-
-	/**
-	 * Focuses the previous status bar entry. If none focused, focuses the last.
-	 */
-	focusPreviousEntry(): void;
-
-	/**
-	 *	Returns true if a status bar entry is focused.
-	 */
-	isEntryFocused(): boolean;
-
-	/**
-	 * Temporarily override statusbar style.
-	 */
-	overrideStyle(style: IStatusbarStyleOverride): IDisposable;
 }
 
 export interface IStatusbarEntryAccessor extends IDisposable {
